QUINTO EJERCICIO REALIZADO Y RESUELTO

+++ Ejercicio03/Ejercicio03/src/ejercicio03/Ejercicio03.java
@@ -8,7 +8,7 @@ public class Ejercicio03 {
         int varl5 = 1;
         int varl6 = 11;
         //FORMLA A CONTINUACION 
-        double resultado = Math.pow(varl1, varl2) + Math.sqrt(varl3) + (double)varl4/(varl5 + varl6);
+        double resultado = Math.pow(varl1, varl2) + Math.sqrt(varl3) + varl4/(varl5 + varl6);
         System.out.println("El resultado es: " + resultado);
         //CORREGIDO Y REALIZADO POR ALEJANDRO JIRON.
